Milestone: make calls for multiple sizes
This will make calls for 900/250, 970/250 and 728/90 sizes, making appnexus requests against multiple placement IDs.

The result is that prebid selects the highest CPM result regardless of which one returns first or size. I can't get the requisite ads to display - they're all giving me the Guardian holiday ads now - but I can tell the right values are being sent up on the request to securepubads - the size of the winning bid is put on the prev_scp URI parameter.